requirements.txt
python-nmap
sslyze
requests
beautifulsoup4
jinja2
# For GUI: Tkinter (usually included with Python, else install python3-tk)
